golf added
it's basically code golf, but for a bunch more things (e.g. get to end of a level in Kirby's Adventure in as few inputs as possible — you could do speedrun strats, or do ACE, etc; also vim race, get to a particular page of a site, etc). Think speedrun challenges with a puzzle flair!